Add test.

+++ 02_sourceCode/01.BackEnd/01.Java/IdentityManagement/src/main/java/app/constant/Constants.java
@@ -13,7 +13,7 @@ public final class Constants {
     public static final String COOKIE_NAME_CSRF = "XSRF-TOKEN";
     public static final String CMD_PYTHON = "python2";
     public static final String FILE_PYTHON_CARD_READ = "/opt/get-card-id.py";
-    public static final int MILLISECOUND_WAIT_CMD = 3000;
+    public static final long MILLISECOUND_WAIT_CMD = 3000;
 
     private Constants() {
     }
